fix(obs-001): move bot_detection out of middleware/ package — collision with middleware.py
Original commit aea083d4 created backend/middleware/ package, which conflicted
with the existing backend/middleware.py module file (CorrelationIDMiddleware,
request_id_var, etc.). Python prefers the package, so all `from middleware import X`
in 100+ test files raised ImportError on CI.

Fix:
- backend/middleware/bot_detection.py -> backend/bot_detection.py (standalone)
- Removed empty backend/middleware/ package (deleted __init__.py)
- Updated import in backend/rate_limiter.py (lazy import inside function)
- Updated import in backend/tests/test_rate_limiter_bot_tier.py

Verification:
- pytest tests/test_rate_limiter_bot_tier.py -> 26 passed in 3.27s
- backend/middleware.py untouched (preserves existing exports)
